+++ changelog.md
@@ -1,3 +1,7 @@
+- **08th August, 2023 - Bernode**
+  - Worked on custom color config component
+- **7th August, 2023 - Eric McWinNEr 😴**
+  - Fixed bug with inline file picker button mode having empty string
 - **5th August, 2023 - Eric McWinNEr 😴**
   - Updated the dropdown and fixed the bug where the dropdown opens when the value changes
   - Fixed bug with currency input not working as expected